requirements.txt
apache-airflow==3.0.0
apache-airflow-providers-google
nba_api>=1.5.2
pandas>=2.2.0
pyarrow>=16.0.0
google-cloud-bigquery>=3.20.0
pandas-gbq>=0.22.0
great_expectations>=1.2.0
streamlit>=1.35.0
plotly>=5.22.0
pytest>=8.2.0
pytest-cov>=5.0.0
ruff>=0.4.0
mypy>=1.10.0
bandit>=1.7.0
